Upload initial project files
This initial upload includes index.html, a pages directory, and a
links and images directory. A png of a pug is in the images directory
and a separate "about.html" page is in the pages directory. These are
subject to change over the next few updates based on the needs of the
project.